[OT-98] [REFACTOR]: Types 폴더 구조 정리
