+++ src/domains/mypage/components/dashboard/DashboardContentList.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Pie } from "react-chartjs-2";
 import { Chart as ChartJS, ArcElement, Tooltip, Legend, ChartOptions } from "chart.js";
 import ChartDataLabels from "chartjs-plugin-datalabels";
-import { DashboardData, TagDetail } from "@/types/dashboard";
+import { DashboardData, TagDetail } from "@/domains/mypage/types/dashboard";
 import TagStatsModal from "@/components/mypage/dashboard/TagStatsModal";
 
 ChartJS.register(ArcElement, Tooltip, Legend, ChartDataLabels);
