+++ src/domains/mypage/components/recenthistory/RecentContentList.tsx
@@ -4,7 +4,7 @@ import { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import RightListScroll from "@/components/mypage/RightListScroll";
 import LeftListScroll from "@/components/mypage/LeftListScroll";
-import { RecentItem } from "@/types/recenthistory";
+import { RecentItem } from "@/domains/mypage/types/recenthistory";
 
 interface RecentContentListProps {
   items: RecentItem[];
