+++ src/domains/mypage/types/dashboard.ts
