+++ src/domains/mypage/types/recenthistory.ts
