+++ src/domains/mypage/types/withdrawcontent.ts
