+++ src/domains/video-contents/components/ContentsContainer.tsx
@@ -9,7 +9,7 @@ import {
   Recommendation,
   SeriesContent,
   SingleContent,
-} from "@/types/contents";
+} from "@/domains/video-contents/types/contents";
 
 interface ContentsContainerProps {
   content: SingleContent | SeriesContent;
