+++ src/domains/video-contents/types/contents.ts
